Some stuff
1. Merging user_info.py from main directory with the one inside microservices
2. Adding haversine wrapper function into food_info,py
3. Adding dockerfile and requirements.txt into microservice
