+++ food_info/requirements.txt
@@ -1,0 +1,5 @@
+Flask==2.2.2
+Flask-Cors==3.0.10
+Flask-SQLAlchemy==3.0.2
+mysql-connector-python==8.0.32
+haversine==2.8.0
