--- location_info/location_info.py
@@ -1,1 +1,0 @@
-# this is to test the branch
