+++ microservices/user_info/user_info.py
@@ -196,7 +196,59 @@ def create_user(name):
         }
     ), 201
 
+'''
+Details for wrapper function below
 
+Function: search for users where the food post latlng is within the users' travel appetite
+Input: food post JSON object
+Output: array of user JSON objects who fulfill the criteria
+'''
+@app.route("/filter_user", methods=['GET'])
+# search for users that are within the distance
+def filter_user():
+    # check input format and data is JSON
+    if request.is_json:
+        try:
+            # get query info
+            query = request.get_json()
+            print("\nReceived an order in JSON:", query)
+
+            # do the actual checking
+            # return list of user objects from userdB
+            all_user_info = user_info.query.all()
+            filtered_users = []
+            if len(all_user_info):
+                # filter for users who are "close" to post according to their travel appetite
+                for user in all_user_info:
+                    user_latitude = user.latitude
+                    user_longitude = user.longitude
+                    user_travel_appetite = float(user.travel_appetite)
+                    query_latitude = query['latitude']
+                    query_longitude = query['longitude']
+                    distance = hs.haversine((user_latitude,user_longitude),(query_latitude, query_longitude))
+
+                    if distance <= user_travel_appetite:
+                        filtered_users.append(user)
+                # return list of user objects where the post is within the user's travel appetite
+                return jsonify(
+                    {
+                        "code": 200,
+                        "data": {
+                            "user": [info.json() for info in filtered_users]
+                        }
+                    }
+                )
+            
+            else:
+                # the else comes here
+                return jsonify(
+                    {
+                        "code": 404,
+                        "message": "No information to be displayed."
+                    }
+                ), 404
+        except:
+            pass
 
 if __name__ == '__main__':
     app.run(port=1111, debug=True)
